Django_Server/requirements.txt
asgiref==3.3.4
Django==3.2.2
install==1.3.4
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
tensorflow==2.3
numpy
pandas
djangorestframework
opencv-python
dlib
PIL
matplotlib
imutils
